Break SearchFilter into multiple files
